[Ref] Silence warnings.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@24785 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ soundlib/Load_mdl.cpp
@@ -654,7 +654,7 @@ bool CSoundFile::ReadMDL(FileReader &file, ModLoadingFlags loadFlags)
 				CopyEnvelope(mptIns->VolEnv, sampleHeader.volEnvFlags, volEnvs);
 				CopyEnvelope(mptIns->PanEnv, sampleHeader.panEnvFlags, panEnvs);
 				CopyEnvelope(mptIns->PitchEnv, sampleHeader.freqEnvFlags, pitchEnvs);
-				mptIns->nFadeOut = (sampleHeader.fadeout + 1u) / 2u;
+				mptIns->nFadeOut = static_cast<uint16>((sampleHeader.fadeout + 1u) / 2u);
 #ifdef MODPLUG_TRACKER
 				if((mptIns->VolEnv.dwFlags & (ENV_ENABLED | ENV_LOOP)) == ENV_ENABLED)
 				{
